requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
DateTime==5.4
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
urllib3==2.2.0
Werkzeug==3.0.1
zope.interface==6.2
